requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
astroid==2.11.7
backports.zoneinfo==0.2.1
black==22.6.0
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
cssselect2==0.7.0
dill==0.3.5.1
distlib==0.3.5
dj-rest-auth==2.2.5
Django==4.1.1
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.7.1
flake8==5.0.4
fonttools==4.37.2
future==0.18.2
html5lib==1.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.9.1
mccabe==0.7.0
mypy-extensions==0.4.3
oscrypto==1.3.0
pathspec==0.9.0
pdfkit==1.0.0
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.9.1
pycparser==2.21
pydyf==0.3.0
pyflakes==2.5.0
pyHanko==0.14.0
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
pylint==2.14.5
PyPDF3==1.0.6
pyphen==0.13.0
python-bidi==0.4.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.4
tqdm==4.64.1
typing-extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
virtualenv==20.16.3
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.8
zopfli==0.2.1